Renaming + removing logs
